==> scripts/01.2_Extract control variables.R <==

library(tidyverse)

alspac <- readRDS("alspac.rds")

#Combine ID and pregnancy ID to generate unique ID 
alspac <- alspac %>% mutate(uniqid = paste0(cidB4083, qlet), collapse="")

#Compute analysis baseline variable (cases productive at the first point in time of the model)

alspac %>% count(kr991a) %>% print(n = 100)

alspac <- alspac %>%
  mutate(baseline_age_7y9m = case_when(kr991a >= -9 ~ 1, TRUE ~ 0))

alspac %>% count(baseline_age_7y9m, kr991a) %>% print(n = 100)

alspac %>% count(baseline_age_7y9m)


alspac %>% count(baseline_age_7y9m, kr882)
alspac %>% count(baseline_age_7y9m, kr881)

#select relevant control variables
alspac_control <- alspac %>% select(uniqid,
                                    baseline_age_7y9m,
                                    a520, 
                                    a525, 
                                    c645, 
                                    c710, 
                                    c711, 
                                    c712, 
                                    c713, 
                                    c716, 
                                    c718, 
                                    c719, 
                                    c721, 
                                    c730, 
                                    c731, 
                                    c732, 
                                    c733, 
                                    c736, 
                                    c738, 
                                    c739, 
                                    c741, 
                                    c804, 
                                    ccs2000, 
                                    ccs2001, 
                                    ccs2030, 
                                    ccs2031, 
                                    f460, 
                                    f500, 
                                    f565, 
                                    f8fp130, 
                                    fdfs130, 
                                    fh9140, 
                                    g517, 
                                    g590, 
                                    h386, 
                                    h480, 
                                    j350, 
                                    j370, 
                                    j556b, 
                                    j556c, 
                                    j556d, 
                                    j600, 
                                    j632, 
                                    kaimd2000q5, 
                                    kaimd2004q5, 
                                    kaimd2007q5, 
                                    kaimd2010q5, 
                                    kaimd2015q5, 
                                    kb011, 
                                    kbimd2000q5, 
                                    kbimd2004q5, 
                                    kbimd2007q5, 
                                    kbimd2010q5, 
                                    kbimd2015q5, 
                                    kcimd2000q5, 
                                    kcimd2004q5, 
                                    kcimd2007q5, 
                                    kcimd2010q5, 
                                    kcimd2015q5, 
                                    kd010, 
                                    kd502a, 
                                    kd502b, 
                                    kd513a, 
                                    kd513b, 
                                    kdimd2000q5, 
                                    kdimd2004q5, 
                                    kdimd2007q5, 
                                    kdimd2010q5, 
                                    kdimd2015q5, 
                                    keimd2000q5, 
                                    keimd2004q5, 
                                    keimd2007q5, 
                                    keimd2010q5, 
                                    keimd2015q5, 
                                    kf010, 
                                    kf452, 
                                    kf452a, 
                                    kf463, 
                                    kf463a, 
                                    kfimd2000q5, 
                                    kfimd2004q5, 
                                    kfimd2007q5, 
                                    kfimd2010q5, 
                                    kfimd2015q5, 
                                    kj010, 
                                    kj462, 
                                    kj462a, 
                                    kj473, 
                                    kj473a, 
                                    kl010, 
                                    kl472, 
                                    kl483, 
                                    kl484, 
                                    kn1000, 
                                    kn4002, 
                                    kn4013, 
                                    kn4014, 
                                    kq010, 
                                    kq348b, 
                                    kq348c, 
                                    kq348d, 
                                    kr010, 
                                    ks1000, 
                                    kt5002, 
                                    kt5014, 
                                    ku706b, 
                                    ku707b, 
                                    ku708b, 
                                    kv1000, 
                                    kw1000, 
                                    kw6601b, 
                                    kw6602b, 
                                    kw6603b, 
                                    kz021, 
                                    l6000, 
                                    l6080, 
                                    m3040, 
                                    m5000, 
                                    n3000, 
                                    n4100, 
                                    n4101, 
                                    n4102, 
                                    n4103, 
                                    n4104, 
                                    n4106, 
                                    n4112, 
                                    n4120, 
                                    n4121, 
                                    n4122, 
                                    n4123, 
                                    n4124, 
                                    n4126, 
                                    n4132, 
                                    n8040, 
                                    n8365a, 
                                    n8365b, 
                                    n8365c, 
                                    p3000, 
                                    p3080, 
                                    pl4100, 
                                    pl4101, 
                                    pl4102, 
                                    pl4103, 
                                    pl4104, 
                                    pl4106, 
                                    pl4112, 
                                    pl4120, 
                                    pl4121, 
                                    pl4122, 
                                    pl4123, 
                                    pl4124, 
                                    pl4126, 
                                    pl4132, 
                                    pm3080, 
                                    pp2002, 
                                    pp2017, 
                                    pp2018, 
                                    pp2019, 
                                    pp2020, 
                                    pp2021, 
                                    pp2022, 
                                    pq3080, 
                                    q3040, 
                                    r2002, 
                                    r2017, 
                                    r2018, 
                                    r2019, 
                                    r2020, 
                                    r2021, 
                                    r2022, 
                                    s3000, 
                                    s3080, 
                                    sa030, 
                                    sa037, 
                                    se030, 
                                    se037, 
                                    t1050, 
                                    ta7025a, 
                                    ta7025b, 
                                    ta7025c, 
                                    tb1000, 
                                    tc4025a, 
                                    tc4025b, 
                                    tc4025c, 
                                    tf1imd2000q5, 
                                    tf1imd2004q5, 
                                    tf1imd2007q5, 
                                    tf1imd2010q5, 
                                    tf1imd2015q5, 
                                    YPB6010, 
                                    YPH2010, 
                                    YPH2011, 
                                    YPH2012, 
                                    b_rnssec_analytical_m, 
                                    c_rnssec_analytical_m, 
                                    f_rnssec_analytical_m, 
                                    g_rnssec_analytical_m, 
                                    h_rnssec_analytical_m, 
                                    j_rnssec_analytical_m, 
                                    c_rnssec_analytical_ptnr_pp, 
                                    f_rnssec_analytical_ptnr_pp, 
                                    g_rnssec_analytical_ptnr_pp, 
                                    h_rnssec_analytical_ptnr_pp, 
                                    j_rnssec_analytical_ptnr_pp, 
                                    pa_rnssec_analytical_p, 
                                    pb_rnssec_analytical_p, 
                                    pd_rnssec_analytical_p, 
                                    pe_rnssec_analytical_p, 
                                    pl_rnssec_analytical_p,
                                    d171a,
                                    pa191a,
                                    kq653,
                                    fefs100)

glimpse(alspac_control)

saveRDS(alspac_control, file = "alspac_control_variables.rds")
